// packages/test/integration_tests/complex_scenarios/handles_complex_nested_structure_with_unions_and_optional_properties/types.ts
export type SubscriptionTier = 'free' | 'basic' | 'business' | undefined;

export interface User {
  id: string;
  avatar: {
    url: string;
  };
  birthday?: Date;
  email: string;
  firstName: string;
  lastName: string;
  subscriptionTier: SubscriptionTier;
  card: {
    currencyCode: string;
  };
  friends: Array<{id: string}>;
}